Clarify rationale for (R, s) signature variant
The main potential advantage of the (e, s) variant is that when using
short hashes, signatures can be encoded using only 48 bytes. However,
the security of these signatures relies on stronger assumptions.

Assuming a more conservative choice of full sized hashes and therefore
64 bytes needed for either variant leaves batch verification as the
decisive factor, clarifying the choice of the (R, s) variant.